airflow/requirements.txt
pymongo==3.10.1
SQLAlchemy
psycopg2-binary
vaderSentiment
